Reconfigure workflow to account for file structure
Move set up into run_mod function so that the only things that need to be modified for each run are the config file, where the models are set up, and the script where run_mod is called.  This requried moving some information from the configuration files into the fun_mod function.